% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{analyse_mecc_cond}
\alias{analyse_mecc_cond}
\title{Perform weighted analysis of (M)ECC data using the conditional approach}
\usage{
analyse_mecc_cond(
  y_name,
  x_formula,
  set_id_name,
  surv,
  surv_tau,
  mecc,
  lower = -10,
  upper = 10
)
}
\arguments{
\item{y_name}{A string indicating cases and controls in \code{mecc}. Note
that this is not the original indicator for event/censoring.}

\item{x_formula}{A \code{formula} object specifying the model assumed for the
covariates, starting with \code{~} (but does not include the outcome
variable).}

\item{set_id_name}{A string indicating the ID of matched sets in the (M)ECC
data. See Details.}

\item{surv}{Estimated baseline survival probability for each subject in
\code{mecc}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}. See Details.}

\item{surv_tau}{Estimated baseline survival probability evaluate at time
\code{tau}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}. See Details.}

\item{mecc}{(M)ECC data. A \code{data.frame}. Make sure the covariates in
\code{x_formula} are all centred at cohort average.}

\item{lower, upper}{Parameters of the function \code{\link{optim}}, which
specifies the lower and upper bounds of the estimated coefficient for
method "Brent" if there is only one coefficient.}
}
\value{
Returns a list consisting of two components:
\describe{
  \item{coef_mecc}{A \code{data.frame} containing the name of covariates in
  the regression model (\code{var}), estimated regression coefficients
  (\code{est}), estimated hazard ratios (\code{exp_est = exp(est)}) standard
  error of the estimates (\code{se}), and the p-values (\code{pval}).}
  \item{optim_obj}{The \code{optim} output from maximising the log-likelihood
  in the weighted analysis.}
}
}
\description{
Perform weighted analysis of (M)ECC data using the conditional approach
}
\details{
This function uses the function \code{\link{optim}} to estimate the
  regression coefficients and the Hessian matrix. The default method of
  \code{\link{optim}} (i.e., "Nelder-Mead") is used if there are more than 1
  coefficient to estimate, otherwise the "Brent" method is used, with the
  search range given by \code{lower} and \code{upper}.
}
\examples{
# Load cohort data
data(cohort_1) 
head(cohort_1)
# Draw simple 1:2 more extreme case-control sample, matched on gender. 
# Let cases be subjects who had the event within 5 years, and controls be 
# selected from those who did not have the event until the 15-th year.
set.seed(1)
dat_mecc <- draw_mecc(cohort = cohort_1, tau0 = 5, tau = 15, 
                      id_name = "id", t_name = "t", delta_name = "y", 
                      match_var_names = "gender", n_per_case = 2)
head(dat_mecc)
# To estimate the HR of age from MECC sample using the weighted approach, 
# it is necessary to center age at the cohort average:
dat_mecc$age_c <- dat_mecc$age - mean(cohort_1$age)
result_mecc <- analyse_mecc_cond(
  y_name = "y_mecc", x_formula = ~ age_c, set_id_name = "set_id_mecc", 
  surv = dat_mecc$surv, surv_tau = dat_mecc$surv_tau, mecc = dat_mecc, 
  lower = -1, upper = 1
)
round(result_mecc$coef_mat[, -1], 3)
# Compare with the estimate from the full cohort:
library(survival)
result_cohort <- summary(coxph(Surv(t, y) ~ age + gender, data = cohort_1))$coef
round(result_cohort["age", ], 3)
# The MECC sample may also be analysed using a logistic regression to 
# estimate the OR of age, which tends to overestimate the HR:
result_logit <- summary(glm(y_mecc ~ age + gender, data = dat_mecc))$coef
round(result_logit, 3)
}
\references{
\itemize{
  \item{Salim A, Ma X, Fall K, et al. Analysis of incidence and prognosis 
  from 'extreme' case–control designs. Stat Med 2014; 33: 5388–5398.}
  \item{Støer NC, Salim A, Bokenberger K, et al. Is the matched extreme
  case–control design more powerful than the nested case–control design?.
  Stat Methods Med Res 2019; 28(6): 1911-1923.}
}
}
\seealso{
\code{\link{draw_mecc}}, \code{\link{llh_mecc_cond}}
}
\author{
Yilin Ning, Nathalie C Støer
}
