% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{analyse_mecc_matched}
\alias{analyse_mecc_matched}
\title{Perform weighted analysis of matched (M)ECC data}
\usage{
analyse_mecc_matched(y_name, x_formula, set_id_name, surv, surv_tau, mecc)
}
\arguments{
\item{y_name}{A string indicating cases and controls in \code{mecc}. Note
that this is not the original indicator for event/censoring.}

\item{x_formula}{A \code{formula} object specifying the model assumed for the
covariates, starting with \code{~} (but does not include the outcome).}

\item{set_id_name}{A string indicating the ID of matched sets in the (M)ECC
data.}

\item{surv}{Estimated baseline survival probability for each subject in
\code{mecc}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}.}

\item{surv_tau}{Estimated baseline survival probability evaluate at time
\code{tau}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}.}

\item{mecc}{(M)ECC data. A \code{data.frame}. Make sure the covariates in
\code{x_formula} are all centred at cohort average.}
}
\value{
Returns a list consisting of two components:
\describe{
  \item{coef_mecc}{A \code{data.frame} containing the name of covariates in
  the regression model (\code{var}), estimated regression coefficients
  (\code{est}), estimated hazard ratios (\code{exp_est = exp(est)}) standard
  error of the estimates (\code{se}), and the p-values (\code{pval}).}
  \item{optim_obj}{The \code{optim} output from maximising the log-likelihood
  in the weighted analysis.}
}
}
\description{
Perform weighted analysis of matched (M)ECC data
}
