% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{assign_kmw0}
\alias{assign_kmw0}
\title{<Private function> Assign appropriate KM-type weight to unique subjects in NCC}
\usage{
assign_kmw0(ncc_nodup, risk_table)
}
\arguments{
\item{ncc_nodup}{NCC data prepared using \code{\link{prepare_cohort}}
function. This data should not include the ID of each matched set or the
time of event in each set.}

\item{risk_table}{Output from \code{\link{compute_risk_tb}}, with an additional
column \code{p}.}
}
\value{
Returns \code{ncc_nodup} with two additional columns: \code{.km_prob}
  for KM-type probability, and \code{.km_weight} for KM-type weight.
}
\description{
<Private function> Assign appropriate KM-type weight to unique subjects in NCC
}
