% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{change_km_names}
\alias{change_km_names}
\title{<private function> Change variable names corresponding to KM-type probability
and weight}
\usage{
change_km_names(dat, km_names)
}
\arguments{
\item{dat}{Cohort or NCC data generated from \code{\link{compute_km_weights}}
or \code{\link{compute_km_weights_controls}}.}

\item{km_names}{Names for KM-type probability and weight, corresponding to the 
last two columns in \code{dat}.}
}
\value{
Returns \code{dat} with names changed for the last two columns.
}
\description{
<private function> Change variable names corresponding to KM-type probability
and weight
}
