% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{change_mecc_names}
\alias{change_mecc_names}
\title{<private function> Change variable names in MECC sample}
\usage{
change_mecc_names(dat, mecc_names)
}
\arguments{
\item{dat}{MECC data generated from \code{\link{draw_mecc}}.}

\item{mecc_names}{Names to assign to the following four variables (in
sequence): the indicator for extreme events, the estimated survival
probability at each time point, the estimated survival probability at
\code{tau}, and the ID for matched sets (see Details). These four variables 
correspond to the last four columns of the MECC sample.}
}
\value{
Returns \code{dat} with names changed for the outcome and set id columns.
}
\description{
<private function> Change variable names in MECC sample
}
