% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohort_2}
\alias{cohort_2}
\title{Simulated cohort 2}
\format{
A data frame with 100000 rows and 8 variables:
\describe{
  \item{id}{Subject ID of each subject.}
  \item{y}{Event/censoring indicator. Events are indicated by y=1.}
  \item{t}{Event/censoring time (in years). Maximum follow-up time is 25 years.}
  \item{x}{Binary exposure.}
  \item{age}{Age of subjects (rounded to integers and mean-centred).}
  \item{age_cat}{Age category of subjects.}
  \item{gender}{Gender of subjects (1 for male and 0 for female).}
  \item{z}{Binary effect modifier.}
}
}
\usage{
cohort_2
}
\description{
Survival outcome was simulated with true hazard: 
\deqn{\log \{h(t)\} = \log \{h_0\}
+ \log(1.5)x + \log(4)z + \log(2)xz + \log(1.01)Gender + \log(1.01)Age.}
}
\examples{
library(survival)
data("cohort_2", package = "SamplingDesignTools")
summary(coxph(Surv(t, y) ~ x * z + age + gender, data = cohort_2))

}
\keyword{datasets}
