% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{compute_kmw0}
\alias{compute_kmw0}
\title{<Private function> Compute KM-type weight from risk table, provided t=0 for all}
\usage{
compute_kmw0(risk_table)
}
\arguments{
\item{risk_table}{Output from \code{\link{compute_risk_tb}}, with an additional
column \code{p}.}
}
\value{
Returns a \code{data.frame} with \code{t_event}, \code{km_weight} and
  matching variables (if any).
}
\description{
<Private function> Compute KM-type weight from risk table, provided t=0 for all
}
