% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{compute_kmw_cohort}
\alias{compute_kmw_cohort}
\title{<Private function> Compute KM-type weights for NCC sample given full cohort,
possibly with dropped controls}
\usage{
compute_kmw_cohort(
  cohort,
  t_start_name = NULL,
  t_name,
  sample_stat,
  match_var_names = NULL,
  n_per_case,
  n_kept = n_per_case,
  attach_weight = TRUE
)
}
\arguments{
\item{cohort}{Cohort data with at least the following information on each
subject: start time (if not 0 for all subjects) and end time of follow-up,
censoring status and matching variables (if any). A \code{data.frame} or a
matrix with column names.}

\item{t_start_name}{Name of the variable in \code{cohort_skeleton} for the
start time of follow-up. A \code{string}. Default is \code{NULL}, where all 
subjects started the follow-up at time 0.}

\item{t_name}{Name of the variable in \code{cohort} for the event or
censoring time. A \code{string}.}

\item{sample_stat}{A numeric vector containing sampling and status
information for each subject in \code{cohort}: use 0 for non-sampled
controls, 1 for sampled (and kept) controls, and integers >=2 for events.
The length of this vector must be the same as the number of rows in
\code{cohort}.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} used
when drawing the NCC. A \code{string} vector. Default is \code{NULL}, i.e.,
the NCC was only time-matched.}

\item{n_per_case}{Number of controls matched to each case.}

\item{n_kept}{Number of sampled controls in each set that were kept in the
final NCC. Default is \code{n_per_case}.}

\item{attach_weight}{Whether to return the subcohort of sampled subjects with
the appropriate KM-type probability and weight attached to each subject
(when \code{attach_weight = TRUE}, the default), or to return a table of
probabilities (and weights) associated with each unique event time. See
Detail.}
}
\description{
<Private function> Compute KM-type weights for NCC sample given full cohort,
possibly with dropped controls
}
\details{
If \code{TRUE} (default), returns a \code{data.frame} containing all
  the unique subjects selected in the NCC sample, with a column for the
  KM-type weight associated with each subject. If \code{FALSE} and all
  subjects started follow-up at time 0 (i.e., \code{t_start_name} is
  \code{NULL}), returns a \code{data.frame} containing the KM-type
  probabilities (\code{.km_prob}) and weights (\code{.km_weight}) for
  controls that were available to be sampled up until each event time. If
  \code{FALSE} and \code{t_start_name} is no \code{NULL}, returns the
  probability of not being sampled at each event time.
}
