% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{compute_risk_tb}
\alias{compute_risk_tb}
\title{<Private function> Compute number at risk at each event time from cohort}
\usage{
compute_risk_tb(cohort, match_var_names)
}
\arguments{
\item{cohort}{Cohort data prepared by \code{\link{prepare_cohort}}.}

\item{match_var_names}{Name(s) of the match variable(s).}
}
\value{
Returns a data.frame with the following columns:
\describe{
  \item{t_event}{Unique event times in \code{cohort_skeleton}.}
  \item{n_event}{Number of events at each event time.}
  \item{n_at_risk}{Number of subjects at risk at each event time.}
  \item{strata}{Strata defined by matching variables. \code{match_var=1} if the
NCC is only time-matched.}
  \item{<each matching variable>}{If the NCC is matched on additional
confounders, each matching variable will be included as a column to the right
of \code{strata}.}
}
}
\description{
<Private function> Compute number at risk at each event time from cohort
}
