% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mcc.R
\name{draw_mcc}
\alias{draw_mcc}
\title{Draw a (matched) case-control (CC) sample from a cohort (or cross-sectional data)}
\usage{
draw_mcc(
  cohort,
  y_name,
  n_per_case,
  match_var_names = NULL,
  weight_name = "ip_weight"
)
}
\arguments{
\item{cohort}{Cohort (or cross-sectional) data. A \code{data.frame} or a
matrix with column names.}

\item{y_name}{Name of the outcome variable in \code{cohort}. Cases are
required to be indicated by integer \code{1}. It is recommended to code the 
outcome variable as an integer vector, where 1=case and 0=non-case.}

\item{n_per_case}{Number of controls to draw for each case.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} used
when drawing a matched CC. A \code{string} vector. Default is \code{NULL},
i.e., the controls will be sampled randomly.}

\item{weight_name}{Name to assign to the column for the inverse probability
weight, which will be created as the last column in the (matched) CC
data returned. Default is \code{"ip_weight"}.}
}
\value{
Returns a \code{data.frame} of the (matched) CC sample, with an 
  additional column for the inverse probability weight for each subject, i.e., 
  1 for cases and the inverse probability of the sampling fraction for each 
  control.
}
\description{
Draw a (matched) case-control (CC) sample from a cohort (or cross-sectional data)
}
\examples{
# Load cohort data
data(cohort_2) # Ignoring that we have the event time in this simulated cohort
head(cohort_2)
# Draw simple 1:2 case-control sample (i.e., with randomly selected controls):
dat_scc <- draw_mcc(cohort = cohort_2, y_name = "y", n_per_case = 2)
head(dat_scc)
# Draw simple 1:2 case-control sample, matched on age group and gender:
dat_mcc <- draw_mcc(cohort = cohort_2, y_name = "y", n_per_case = 2,
                    match_var_names = c("age_cat", "gender"),
                    weight_name = ".w")
head(dat_mcc)
}
\author{
Yilin Ning, Anastasia Lam, Marie Reilly
}
