% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mcc.R
\name{draw_mcc0}
\alias{draw_mcc0}
\title{Private function: draw matched case-control data from cohort, taking all 
cases available}
\usage{
draw_mcc0(cohort, y_name, n_per_case, match_var_names)
}
\description{
Private function: draw matched case-control data from cohort, taking all 
cases available
}
