% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{draw_mecc}
\alias{draw_mecc}
\title{Draw matched (more) extreme case-control samples}
\usage{
draw_mecc(
  cohort,
  tau0,
  tau,
  id_name,
  t_name,
  delta_name,
  match_var_names = NULL,
  n_per_case,
  mecc_names = c("y_mecc", "surv", "surv_tau", "set_id_mecc")
)
}
\arguments{
\item{cohort}{Cohort data.}

\item{tau0}{Only subjects who had event before \code{tau0} are considered as
cases.}

\item{tau}{Only subjects who did not have event at or before \code{tau} are
eligible to be selected as controls. Extreme case-control (ECC) has 
\code{tau = tau0}, and more extreme case-control (MECC) has 
\code{tau > tau0}.}

\item{id_name}{Name of subject ID.}

\item{t_name}{Name of time variable.}

\item{delta_name}{Name of event/censoring indicator.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} used
when drawing a matched MECC sample. A \code{string} vector. Default is
\code{NULL}, i.e., the controls will be sampled randomly.}

\item{n_per_case}{Number of controls to match to each case.}

\item{mecc_names}{Names to assign to the following four variables (in
sequence): the indicator for extreme events, the estimated survival
probability at each time point, the estimated survival probability at
\code{tau}, and the ID for matched sets (see Details). Default value is
\code{c("y_mecc", "surv", "surv_tau", "set_id_mecc")}.}
}
\value{
Returns a \code{data.frame} of the (more) extreme case-control
  sample, with four additional columns as described in parameter
  \code{mecc_names}.
}
\description{
Draw matched (more) extreme case-control samples
}
\details{
This function draws a (more) extreme case-control sample from a
  cohort, with or without matching on confounder(s), with \code{n_per_case}
  controls drawn for each case. Eligible cases are subjects who had the event
  before time \code{tau0} (and all of them are sampled), and eligible
  controls are subjects who do not have the event until time \code{tau}. This
  function throws an error if there is not enough controls to select from,
  otherwise it uses the \code{\link{draw_mcc}} function to draw a
  case-control sample. 
  
  This package implements the conditional approach described in Section 2.2
  of Salim et al 2014, where cases and controls are "individually matched".
  Therefore, although controls are selected randomly or by frequency
  matching, \code{n_per_case} controls would be randomly assigned to a case
  to form a "matched set" after drawing the sample.
}
\examples{
library(SamplingDesignTools)
# Load cohort data
data("cohort_1")
head(cohort_1)
# Draw simple 1:2 more extreme case-control sample, matched on gender.
# Let cases be subjects who had the event within 5 years, and controls be
# selected from those who did not have the event until the 15-th year.
dat_mecc <- draw_mecc(cohort = cohort_1, tau0 = 5, tau = 15,
                      id_name = "id", t_name = "t", delta_name = "y",
                      match_var_names = "gender", n_per_case = 2)
head(dat_mecc)
# Note that the new event indicator, `y_mecc`, is different from the original
# event, `y`, in the cohort:
identical(dat_mecc$y, dat_mecc$y_mecc) # Expect FALSE
}
\references{
\itemize{
  \item{Salim A, Ma X, Fall K, et al. Analysis of incidence and prognosis 
  from 'extreme' case–control designs. Stat Med 2014; 33: 5388–5398.}
  \item{Støer NC, Salim A, Bokenberger K, et al. Is the matched extreme
  case–control design more powerful than the nested case–control design?.
  Stat Methods Med Res 2019; 28(6): 1911-1923.}
}
}
\seealso{
\code{\link{llh_mecc_cond}}, \code{\link{analyse_mecc_cond}}
}
\author{
Yilin Ning, Nathalie C Støer
}
