% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{draw_mecc_matched}
\alias{draw_mecc_matched}
\title{Draw matched (more) extreme case-control samples}
\usage{
draw_mecc_matched(
  cohort,
  tau0,
  tau,
  id_name,
  t_name,
  delta_name,
  match_var_names,
  n_per_case
)
}
\arguments{
\item{cohort}{Cohort data.}

\item{tau0}{Only subjects who had event before \code{tau0} are considered as
cases.}

\item{tau}{Only subjects who did not have event at or before \code{tau} are
eligible to be selected as controls. Extreme case-control (ECC) has 
\code{tau = tau0}, and more extreme case-control (MECC) has 
\code{tau > tau0}.}

\item{id_name}{Name of subject ID.}

\item{t_name}{Name of time variable.}

\item{delta_name}{Name of event/censoring indicator.}

\item{match_var_names}{Confounder(s) to match on.}

\item{n_per_case}{Number of controls to match to each case.}
}
\description{
Draw matched (more) extreme case-control samples
}
\author{
Yilin Ning, Nathalie C Stoer
}
