% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_match.R
\name{draw_ncc_cm}
\alias{draw_ncc_cm}
\title{Draw counter-matched NCC from cohort}
\usage{
draw_ncc_cm(
  cohort,
  y_name = NULL,
  t_name = NULL,
  match_var_name = NULL,
  include_var_name = NULL,
  ml = 1
)
}
\arguments{
\item{cohort}{Cohort data. A \code{data.frame} or a matrix with column names.}

\item{y_name}{Name of the variable in \code{cohort} indicating the
event-censoring status of each subject, where 1 indicates events. A
\code{string}. If a vector is supplied, only the first element will be
used.}

\item{t_name}{Name of the variable in \code{cohort} for the time of event or
censoring. A \code{string}. Currently do not support scenarios where
subjects enter the cohort at different time. If a vector is supplied, only
the first element will be used.}

\item{match_var_name}{Name of the categorical variable in \code{cohort} to
count-match on, which can be the exposure or the surrogate for the
exposure. A \code{string}. If a vector is supplied, only the first element
will be used.}

\item{include_var_name}{A string vector containing additional variables in
\code{cohort} to include in the counter-matched NCC sample. Default is
\code{NULL}.}

\item{ml}{Number of subjects to draw from each strata defined by the match
variable (including the case). Default is 1.}
}
\value{
Returns a \code{data.frame} of the counter-matched NCC sample.
}
\description{
Draw counter-matched NCC from cohort
}
\examples{
data(cohort_1)
head(cohort_1)
# Counter-match on binary indicator for age:
cohort_1$age_bin <- as.numeric(cohort_1$age < 50)
ncc_cm_bin <- draw_ncc_cm(cohort = cohort_1, y_name = "y", t_name = "t", 
                          match_var_name = "age_bin", 
                          include_var_name = c("age", "gender"), ml = 1)
head(ncc_cm_bin, 10)
}
