% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{llh_matched}
\alias{llh_matched}
\title{Negative log-likelihood in weighted analysis of matched (M)ECC sample}
\usage{
llh_matched(beta, x_mat, y_name, set_id_name, surv, surv_tau, mecc)
}
\arguments{
\item{beta}{A vector of regression coefficients corresponding to
\code{x_formula}.}

\item{x_mat}{A \code{model.matrix} (without the first column that has value 
1 for all subjects).}

\item{y_name}{A string indicating cases and controls in \code{mecc}. Note
that this is not the original indicator for event/censoring.}

\item{set_id_name}{A string indicating the ID of matched sets in the (M)ECC
data.}

\item{surv}{Estimated baseline survival probability for each subject in
\code{mecc}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}.}

\item{surv_tau}{Estimated baseline survival probability evaluate at time
\code{tau}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}.}

\item{mecc}{(M)ECC data. A \code{data.frame}. Make sure the covariates in
\code{x_formula} are all centred at cohort average.}
}
\description{
Negative log-likelihood in weighted analysis of matched (M)ECC sample
}
\author{
Yilin Ning, Nathalie C Stoer
}
