% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{llh_mecc_cond}
\alias{llh_mecc_cond}
\title{<Private function> Negative log-likelihood in weighted analysis of (M)ECC sample}
\usage{
llh_mecc_cond(beta, x_mat, y_name, set_id_name, surv, surv_tau, mecc)
}
\arguments{
\item{beta}{A vector of regression coefficients corresponding to
\code{x_formula}.}

\item{x_mat}{A \code{model.matrix} (without the first column that has value 
1 for all subjects).}

\item{y_name}{A string indicating cases and controls in \code{mecc}. Note
that this is not the original indicator for event/censoring.}

\item{set_id_name}{A string indicating the ID of matched sets in the (M)ECC
data. See Details.}

\item{surv}{Estimated baseline survival probability for each subject in
\code{mecc}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}. See Details.}

\item{surv_tau}{Estimated baseline survival probability evaluate at time
\code{tau}, based on the underlying cohort. The length of this variable 
must be the same as the number of rows in \code{mecc}. See Details.}

\item{mecc}{(M)ECC data. A \code{data.frame}. Make sure the covariates in
\code{x_formula} are all centred at cohort average.}
}
\value{
Returns the negative log-likelihood in the weighted analysis of a
  (more) extreme case-control sample.
}
\description{
<Private function> Negative log-likelihood in weighted analysis of (M)ECC sample
}
\details{
This function implements the conditional approach described in
  Section 2.2.1 of Salim et al 2014. This requires controls to be
  "individually matched" to each control. If users used the function
  \code{\link{draw_mecc}} to draw the MECC sample, the sample would include a
  variable that indicates the matched sets, and two variables corresponding
  to \code{surv} and \code{surv_tau}. Otherwise users should randomly match
  each case to controls (within strata if the MECC sample is matched on
  confounder(s)) and compute them from the full cohort as described in
  Section 2.2 of Salim et al 2014.
}
\references{
\itemize{
  \item{Salim A, Ma X, Fall K, et al. Analysis of incidence and prognosis 
  from 'extreme' case–control designs. Stat Med 2014; 33: 5388–5398.}
  \item{Støer NC, Salim A, Bokenberger K, et al. Is the matched extreme
  case–control design more powerful than the nested case–control design?.
  Stat Methods Med Res 2019; 28(6): 1911-1923.}
}
}
\seealso{
\code{\link{draw_mecc}}, \code{\link{analyse_mecc_cond}}
}
\author{
Yilin Ning, Nathalie C Støer
}
