% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{match_risk_table}
\alias{match_risk_table}
\title{<Private function> Match number at risk at coarsened time to NCC cases}
\usage{
match_risk_table(
  ncc_cases,
  risk_table_manual,
  t_coarse_name,
  t_name,
  match_var_names = NULL
)
}
\arguments{
\item{ncc_cases}{Cases in NCC data. A \code{data.frame} or a matrix with
column names. This data should not include the ID of each matched set.}

\item{risk_table_manual}{A \code{data.frame} with columns \code{t_event}
(unique event times, possibly coarsened), \code{n_at_risk} (number of
subjects at risk at each \code{t_event} in the underlying cohort), and
additional columns for matching variables, if any. Make sure the matching
variables have the same column names as in \code{ncc_cases} and
\code{match_var_names}.}

\item{t_coarse_name}{Name of the column of event time in each matched set in
\code{ncc}, possibly coarsened to the same level as \code{t_event} in
\code{risk_table_manual}. A \code{string}.}

\item{t_name}{Name of the column of the exact event time. A \code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) in
\code{ncc_cases} used when drawing the NCC. A \code{string} vector.}
}
\value{
Returns a data.frame with the following columns:
\describe{
  \item{t_event}{Unique event times (exact, not coarsened).}
  \item{n_event}{Number of events at each event time.}
  \item{n_at_risk}{Number of subjects at risk at each event time.}
  \item{strata}{Strata defined by matching variables. \code{match_var=1} if the
NCC is only time-matched.}
  \item{<each matching variable>}{If the NCC is matched on additional
confounders, each matching variable will be included as a column to the right
of \code{strata}.}
}
}
\description{
<Private function> Match number at risk at coarsened time to NCC cases
}
