% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_cohorts}
\alias{mini_cohorts}
\alias{mini_cohort}
\alias{mini_cohort2}
\title{Simulated mini cohorts}
\format{
A data frame with 10 rows and the following variables:
\describe{
  \item{id}{Subject ID.}
  \item{t}{Event/censoring time, with follow-up starting at time 0.}
  \item{t_start}{Start time of follow-up.}
  \item{t_end}{End time of follow-up.}
  \item{status}{Censoring status.}
}

An object of class \code{data.frame} with 10 rows and 4 columns.
}
\usage{
mini_cohort

mini_cohort2
}
\description{
Mini cohorts with common and staggered entry time.
}
\keyword{datasets}
