% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_at_risk}
\alias{n_at_risk}
\title{Number of subjects at risk at each time point in \code{ncc_2}}
\format{
A data frame with 2773 rows and 4 variables:
\describe{
  \item{t}{Event time (in years) of cases in \code{ncc_2}.}
  \item{gender}{Gender of subjects (1 for male and 0 for female).}
  \item{age_cat}{Age category of subjects.}
  \item{n.risk}{Number of subjects at risk at each event time in each gender-age strata.}
}
}
\usage{
n_at_risk
}
\description{
Number of subjects at risk at each time point in \code{ncc_2}
}
\details{
Note how column names in \code{n_at_risk} match those in
  \code{ncc_2}. This is critical for computing the KM-type weights for
  subjects in \code{ncc_2}.
}
\keyword{datasets}
