% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncc_1}
\alias{ncc_1}
\title{Nested case-control (NCC) data 1}
\format{
A data frame with 1164 rows and 6 variables:
\describe{
  \item{Set}{Set ID.}
  \item{Map}{Row numbers in \code{cohort_1}.}
  \item{Time}{Event time (in years) of the case in each \code{Set}.}
  \item{Fail}{Case-control indicator.}
  \item{age}{Age of subjects (rounded to integers).}
  \item{gender}{Gender of subjects (1 for male and 0 for female).}
}
}
\usage{
ncc_1
}
\description{
Time-matched data drawn from \code{\link{cohort_1}}, with one control
matched to each case.
}
\keyword{datasets}
