% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncc_2}
\alias{ncc_2}
\title{Nested case-control (NCC) data 2}
\format{
A data frame with 16638 rows and 10 variables:
\describe{
  \item{Set}{Set ID.}
  \item{Map}{Row numbers in \code{cohort_2}.}
  \item{Time}{Event time (in years) of the case in each \code{Set}.}
  \item{Fail}{Case-control indicator.}
  \item{age_cat}{Age category of subjects.}
  \item{gender}{Gender of subjects (1 for male and 0 for female).}
  \item{x}{Binary exposure.}
  \item{age}{Age of subjects (rounded to integers).}
  \item{z}{Binary effect modifier.}
  \item{t}{Actual time/censoring (in years) of each subject.}
}
}
\usage{
ncc_2
}
\description{
NCC data drawn from \code{\link{cohort_2}} matched on age group
  and gender, with five controls matched to each case.
}
\keyword{datasets}
