% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{p_not_sampled}
\alias{p_not_sampled}
\title{<Private function> Compute the probability that an eligible non-case is not
sampled at each event time}
\usage{
p_not_sampled(n_event, n_at_risk, n_per_case)
}
\arguments{
\item{n_event}{A vector of number of event at each event time (1 if no tie in
event time, or >1 in the presence of ties). Non-integers are rounded down.}

\item{n_at_risk}{A vector of the number of subject at risk at each event
time. Non-integers are rounded down.}

\item{n_per_case}{Number of controls to select per case. A positive integer.
Non-integers are rounded down.}
}
\value{
Returns a vector of probabilities.
}
\description{
<Private function> Compute the probability that an eligible non-case is not
sampled at each event time
}
