% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{prepare_cohort}
\alias{prepare_cohort}
\title{<Private function> Prepare the skeleton cohort for subsequent steps.}
\usage{
prepare_cohort(
  cohort,
  t_start_name,
  t_name,
  y_name,
  match_var_names,
  print_message = TRUE
)
}
\arguments{
\item{cohort}{Cohort data with at least the following information on each
subject: start time (if not 0 for all subjects) and end time of follow-up,
censoring status and matching variables (if any). A \code{data.frame} or a
matrix with column names.}

\item{t_start_name}{Name of the variable in \code{cohort} for the start time
of follow-up. A \code{string}.}

\item{t_name}{Name of the variable in \code{cohort} for the event or
censoring time. A \code{string}.}

\item{y_name}{Name of the column of censoring status in each matched set in
\code{cohort}, with 1 for event and 0 for censoring. A \code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} used
when drawing the NCC. A \code{string} vector.}

\item{print_message}{Whether to print the message regarding start and exit
time. Default is \code{TRUE}.}
}
\description{
<Private function> Prepare the skeleton cohort for subsequent steps.
}
