% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{prepare_ncc_cases}
\alias{prepare_ncc_cases}
\title{<Private function> Prepare the NCC cases for subsequent steps.}
\usage{
prepare_ncc_cases(ncc_cases, t_name, match_var_names)
}
\arguments{
\item{ncc_cases}{Cases in NCC data. A \code{data.frame} or a matrix with
column names. This data should not include the ID of each matched set.}

\item{t_name}{Name of the variable in \code{ncc_cases} for event time. A
\code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) in
\code{ncc_cases} used when drawing the NCC. A \code{string} vector.}
}
\description{
<Private function> Prepare the NCC cases for subsequent steps.
}
