% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{prepare_risk_table}
\alias{prepare_risk_table}
\title{Prepare a template risk table given a nested case-control sample}
\usage{
prepare_risk_table(
  ncc,
  t_match_name,
  y_name,
  match_var_names = NULL,
  csv_file = NULL
)
}
\arguments{
\item{ncc}{Nested case-control data. A \code{data.frame} or a matrix with
column names.}

\item{t_match_name}{Name of the column of event time in each matched set in
\code{ncc}. A \code{string}. Time should be recorded in a reasonable
resolution of measurement (e.g., in months or years) such that the number
of subjects at risk in the cohort is available.}

\item{y_name}{Name of the case/control indicator in \code{ncc}, with 1 for
case and 0 for control A \code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) used when drawing the
NCC. A \code{string} vector. Default is NULL, i.e., the NCC was only
time-matched.}

\item{csv_file}{Name of CSV file to write the template to. If left
unspecified, the template will be returned as a \code{data.frame} instead.}
}
\value{
If \code{csv_file} is not supplied, this function returns a
  \code{data.frame} that contains the unique event times, additional matching
  variables (if \code{match_var_names} is supplied), and an empty column
  named \code{n.risk}. Column names in this \code{data.frame} should not be
  modified. If \code{csv_file} is supplied, this function write the
  \code{data.frame} to the designated file instead. Users should fill this
  column with the number of subjects at risk at each time point in the
  stratum, and later use this \code{data.frame} or file to compute the
  KM-type weights for the subjects in the NCC data.
}
\description{
Given a nested case-control (NCC) data, create a template for
  users to fill in the number of subjects at risk in the cohort at each event
  time within each sampling stratum (if applicable).
}
\examples{
library(SamplingDesignTools)
data("ncc_2")
risk_table <- prepare_risk_table(ncc = ncc_2, t_match_name = "Time", y_name = "Fail",
                                 match_var_names = c("age_cat", "gender"))
head(risk_table)
}
