% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_case_control.R
\name{recode_extreme}
\alias{recode_extreme}
\title{<Private function> Recode subjects as cases and potential controls for (M)ECC
sampling}
\usage{
recode_extreme(t, delta, tau0, tau)
}
\arguments{
\item{t}{Event/censoring time.}

\item{delta}{Event/censoring status.}

\item{tau0}{Only subjects who had event before \code{tau0} are considered as
cases.}

\item{tau}{Only subjects who did not have event at or before \code{tau} are
eligible to be selected as controls. Extreme case-control (ECC) has 
\code{tau = tau0}, and more extreme case-control (MECC) has 
\code{tau > tau0}.}
}
\description{
<Private function> Recode subjects as cases and potential controls for (M)ECC
sampling
}
