% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{cohort_2}
\alias{cohort_2}
\title{Simulated cohort 2}
\format{
A data frame with 100000 rows and 8 variables:
\describe{
  \item{id}{Subject ID of each subject.}
  \item{y}{Event/censoring indicator. Events are indicated by y=1.}
  \item{t}{Event/censoring time (in years). Maximum follow-up time is 25 years.}
  \item{x}{Binary exposure.}
  \item{age}{Age of subjects (rounded to integers and mean-centred).}
  \item{age_cat}{Age category of subjects.}
  \item{gender}{Gender of subjects (1 for male and 0 for female).}
  \item{z}{Binary effect modifier.}
}
}
\usage{
cohort_2
}
\description{
Simulated cohort 2
}
\keyword{datasets}
