% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{compute_km_weights}
\alias{compute_km_weights}
\title{Compute Kaplan-Meier type weights for (matched) nested case-control (NCC)
sample}
\usage{
compute_km_weights(
  cohort = NULL,
  ncc = NULL,
  id_name = NULL,
  risk_table_manual = NULL,
  t_start_name = NULL,
  t_name = NULL,
  sample_stat = NULL,
  t_match_name = t_name,
  y_name = NULL,
  match_var_names = NULL,
  n_per_case,
  n_kept = n_per_case,
  return_risk_table = FALSE,
  km_names = c("km_prob", "km_weight")
)
}
\arguments{
\item{cohort}{Cohort data with at least the following information on each
subject: start time (if not 0 for all subjects) and end time of follow-up,
censoring status and matching variables (if any). A \code{data.frame} or a
matrix with column names.}

\item{ncc}{(Matched) NCC data, if \code{cohort} is not available.
\strong{This data should not include the ID of each matched set, but should
include the actual event/censoring time of each subject.} A
\code{data.frame} or a matrix with column names.}

\item{id_name}{Name of the column indicating subject ID in \code{ncc}, if
\code{cohort} is not available.}

\item{risk_table_manual}{Number of subjects at risk at time of each cases in
the NCC, if \code{cohort} is not available. A \code{data.frame} or a matrix
with column names. See Details.}

\item{t_start_name}{Name of the variable in \code{cohort} or \code{ncc} for
the start time of follow-up. A \code{string}. Default is \code{NULL}, i.e.,
every subject started the follow-up at time 0.}

\item{t_name}{Name of the variable in \code{cohort} or \code{ncc} for the
time of event or censoring. A \code{string}. Note that if \code{ncc} is
supplied, in order to correctly compute the weight for each sampled control
this should be the actual time of censoring, not the time of event of the
case in the same matched set.}

\item{sample_stat}{A numeric vector containing sampling and status
information for each subject in \code{cohort}: use 0 for non-sampled
controls, 1 for sampled (and kept) controls, and integers >=2 for events.
The length of this vector must be the same as the number of rows in
\code{cohort}.}

\item{t_match_name}{Name of the column of event time in each matched set in
\code{ncc}, possibly coarsened to the same level as \code{t_event} in
\code{risk_table_manual}. A \code{string}. Default is \code{t_name}, i.e., 
not coarsened.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} or
\code{ncc} used when drawing the NCC. A \code{string} vector. Default is
\code{NULL}, i.e., the NCC was only time-matched. The corresponding column
in \code{risk_table_manual} must have the same name.}

\item{n_per_case}{Number of controls matched to each case.}

\item{return_risk_table}{Whether the risk table should be returned. Default
is \code{FALSE}.}

\item{km_names}{Column names for the KM-type probability (the first element)
and weight (the second element) computed, if these two columns are to be 
attached to each subject in the input data. Default is
\code{c("km_prob", "km_weight")}.}
}
\description{
Compute Kaplan-Meier type weights for (matched) nested case-control (NCC)
sample
}
\details{
When the full cohort is not available, in order to compute the
  correct weights for each sampled control in the NCC sample, it is important
  to keep the actual time of event or censoring of each subject in the NCC
  sample, which should be specified as \code{t_name} in the input. Since the 
  number of subjects in each risk set will be supplied separately (i.e., as 
  \code{n_at_risk}) in such scenario, \code{t_match_name} is required to 
  map each control to the appropriate risk set. \code{t_match_name} may be 
  the same as \code{t_name} if the exact risk set is available in 
  \code{n_at_risk}, but when the full cohort is not available the risk set is 
  usually approximated by using a coarsened version of \code{t_name}. For 
  example, when controls were drawn from a population registry by matching on
  the exact date of death of cases, birth cohort and gender, the number at
  risk may be approximated by using the population size in the year of event
  in the same birth cohort of the same gender. In this scenario
  \code{t_match_name} would be the year of \code{t_name}.
}
\examples{
# Load cohort data
data(cohort_2)
head(cohort_2)
# Load an NCC sampled drawn from cohort_2 using the following code
# ncc_2 <- Epi::ccwc(exit = t, fail = y, controls = 5, 
#                    match = list(age_cat, gender), include = list(x, age, z), 
#                    data = cohort_2)
data(ncc_2)
head(ncc_2)
# Map the NCC sample to the original cohort, break the matching, identify the 
# subjects selected into the NCC, and return this subset with KM type weights 
# computed for them.
# First create the sampling and status indicator:
sample_stat <- numeric(nrow(cohort_2))
sample_stat[unique(ncc_2$Map[ncc_2$Fail == 0])] <- 1
sample_stat[ncc_2$Map[ncc_2$Fail == 1]] <- 2
# Then find the sampled subset and compute weights:
ncc_nodup <- compute_km_weights(cohort = cohort_2, t_name = "t", y_name = "y",
                                match_var_names = c("age_cat", "gender"), 
                                sample_stat = sample_stat, n_per_case = 5)
head(ncc_nodup)
# Alternatively, if the cohort is not available, the weights can be computed 
# as long as number of subjects at risk at event times in each strata is 
# available elsewhere, and the actual time of event/censoring is available 
# for each subject in the NCC.
# Compute the number of subjects at risk from cohort_2:
risk_table <- compute_risk_table(cohort = cohort_2, t_name = "t", y_name = "y",
                                 match_var_names = c("age_cat", "gender"))
head(risk_table)
# he following command computes the same weights as in ncc_nodup:
ncc_nodup_v2 <- compute_km_weights(ncc = ncc_2[, -1], risk_table_manual = risk_table, 
                                   id_name = "Map", t_match_name = "Time", 
                                   t_name = "t", y_name = "Fail", 
                                   match_var_names = c("age_cat", "gender"),
                                   n_per_case = 5)
head(ncc_nodup_v2)
}
\seealso{
\code{\link{compute_risk_table}}
}
