% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{compute_km_weights_controls}
\alias{compute_km_weights_controls}
\title{Compute Kaplan-Meier type weights for newly collected NCC controls}
\usage{
compute_km_weights_controls(
  ncc_controls,
  prob_table_manual,
  t_start_name = NULL,
  t_name,
  match_var_names = NULL
)
}
\arguments{
\item{ncc_controls}{Newly collected NCC controls, where each row corresponds
to a unique subject. This data should include the actual event/censoring
time of each subject. A \code{data.frame} or a matrix with column names.}

\item{prob_table_manual}{Probability of not being sampled at each event time,
computed using function \code{\link{compute_p_not_sampled}}.}

\item{t_start_name}{Name of the variable in \code{ncc_controls} for the start
time of follow-up. A \code{string}. Default is \code{NULL}, i.e., every
subject started the follow-up at time 0.}

\item{t_name}{Name of the variable in \code{ncc_controls} for the time of
event or censoring. A \code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) in
\code{ncc_controls} used when drawing the NCC. A \code{string} vector.
Default is \code{NULL}, i.e., the NCC was only time-matched. The
corresponding column in \code{prob_table_manual} must have the same name.}
}
\description{
Compute Kaplan-Meier type weights for newly collected NCC controls
}
\seealso{
\code{\link{compute_p_not_sampled}}
}
