% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{compute_kmw_ncc}
\alias{compute_kmw_ncc}
\title{<Private function> Compute KM-type weight for NCC sample given information on 
underlying cohort, possibly with dropped controls}
\usage{
compute_kmw_ncc(
  ncc,
  id_name = NULL,
  risk_table_manual,
  t_start_name = NULL,
  t_name,
  t_match_name = t_name,
  y_name,
  match_var_names = NULL,
  n_per_case,
  n_kept = n_per_case,
  return_risk_table = FALSE,
  km_names = c(".km_prob", ".km_weight")
)
}
\arguments{
\item{ncc}{NCC data. A \code{data.frame} or a matrix with column names. This
data should not include the ID of each matched set, but should include the
actual event/censoring time of each subject.}

\item{id_name}{Name of the column of the subject ID. A \code{string}.}

\item{risk_table_manual}{A \code{data.frame} with columns \code{t_event}
(unique event times, possibly coarsened), \code{n_at_risk} (number of
subjects at risk at each \code{t_event} in the underlying cohort), and
additional columns for matching variables, if any. Make sure the matching
variables have the same column names as in \code{ncc_cases} and
\code{match_var_names}.}

\item{t_name}{Name of the column of the exact event/censoring time of each
subject. Note that for controls, this should not be the time of the event
in the same matched set. A \code{string}.}

\item{t_match_name}{Name of the column of event time in each matched set in
\code{ncc}, possibly coarsened to the same level as \code{t_event} in
\code{risk_table_manual}. A \code{string}. Default is \code{t_name}, i.e., 
not coarsened.}

\item{y_name}{Name of the column of censoring status in each matched set in
\code{ncc}, with 1 for event and 0 for censoring. A \code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) in
\code{ncc_cases} used when drawing the NCC. A \code{string} vector.}

\item{n_per_case}{Number of controls matched to each case.}

\item{n_kept}{Number of sampled controls in each set that were kept in the
final NCC. Default is \code{n_per_case}.}

\item{return_risk_table}{Whether the risk table should be returned. Default
is \code{FALSE}.}

\item{km_names}{Column names for the KM-type probability (the first element)
and weight (the second element) computed, if these two columns are to be 
attached to each subject in the input data. Default is
\code{c(".km_prob", ".km_weight")}.}
}
\value{
If \code{return_risk_table = FALSE} (the default), returns a
  \code{data.frame} containing all the unique subjects selected in the NCC
  sample, with a column for the KM-type weight associated with each subject.
  If \code{return_risk_table = TRUE}, returns a list containing this
  subcohort (\code{dat}) and the risk table (\code{risk_table}), which is a
  \code{data.frame} containing the distinct (and exact) event time
  (\code{t_event}), matching variables (if any), and the number of subject at
  risk at each event time in each strata defined by matching variables
  (\code{n_at_risk}).
}
\description{
<Private function> Compute KM-type weight for NCC sample given information on 
underlying cohort, possibly with dropped controls
}
