% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_type_weights.R
\name{compute_risk_table}
\alias{compute_risk_table}
\title{Compute number at risk at each event time from a "skeleton cohort"}
\usage{
compute_risk_table(
  cohort,
  t_start_name = NULL,
  t_name,
  y_name,
  match_var_names = NULL
)
}
\arguments{
\item{cohort}{Cohort data with at least the following information on each
subject: start time (if not 0 for all subjects) and end time of follow-up,
censoring status and matching variables (if any). A \code{data.frame} or a
matrix with column names.}

\item{t_start_name}{Name of the variable in \code{cohort} for the start time
of follow-up. A \code{string}. Default is \code{NULL}, i.e., every subject
started the follow-up at time 0.}

\item{t_name}{Name of the variable in \code{cohort} for the event or
censoring time. A \code{string}.}

\item{y_name}{Name of the column of censoring status in each matched set in
\code{cohort}, with 1 for event and 0 for censoring. A \code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} used
when drawing the NCC. A \code{string} vector. Default is \code{NULL}, i.e.,
the NCC was only time-matched.}
}
\value{
Returns a data.frame with the following columns:
\describe{
  \item{t_event}{Unique event times in \code{cohort_skeleton}.}
  \item{n_event}{Number of events at each event time.}
  \item{n_at_risk}{Number of subjects at risk at each event time.}
  \item{<each matching variable>}{If the NCC is matched on additional
confounders, each matching variable will be included as a column to the right
of \code{strata}.}
}
}
\description{
Compute number at risk at each event time from a "skeleton cohort"
}
\examples{
library(SamplingDesignTools)
# With common entry time:
data("mini_cohort")
mini_cohort
compute_risk_table(cohort = mini_cohort, t_name = "t", y_name = "status")
# With staggered entry time:
data("mini_cohort2")
mini_cohort2
compute_risk_table(cohort = mini_cohort2, t_start_name = "t_start", 
                   t_name = "t_end", y_name = "status")
}
\seealso{
\code{\link{compute_km_weights}}
}
