% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mcc.R
\name{draw_scc0}
\alias{draw_scc0}
\title{Private function: draw simple case-control data from cohort, taking all 
cases available}
\usage{
draw_scc0(cohort, y_name, n_per_case)
}
\description{
Private function: draw simple case-control data from cohort, taking all 
cases available
}
