% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{prep_km1}
\alias{prep_km1}
\title{Assemble KM table and NCC data without duplicates when cohort data is
available}
\usage{
prep_km1(
  cohort,
  t_name,
  y_name,
  sample_stat,
  keep_stat = NULL,
  match_var_names = NULL,
  n_per_case
)
}
\arguments{
\item{cohort}{Cohort data. A \code{data.frame} or a matrix with column names.}

\item{t_name}{Name of the variable in \code{cohort} for the time of event or
censoring. A \code{string}.}

\item{y_name}{Name of the variable in \code{cohort} indicating the
event-censoring status of each subject, where 1 indicates events. A
\code{string}.}

\item{sample_stat}{A numeric vector containing sampling and status
information for each subject in \code{cohort}: use 0 for non-sampled
controls, 1 for sampled controls, and integers >=2 for events. The length 
of this vector must be the same as the number of rows in \code{cohort}.}

\item{keep_stat}{A numeric vector indicating whether each subject in
\code{cohort} are kept in the final NCC: use 1 for subjects who were in the
final NCC, and 0 for subjects who were kept or never selected. The length
of this vector must be the same as the number of rows in \code{cohort}.
When unspecified, the function assumes all subjects are kept in the final
NCC.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} used
when drawing the NCC. A \code{string} vector. Default is \code{NULL}, i.e.,
the NCC was only time-matched.}

\item{n_per_case}{Number of controls matched to each case.}
}
\description{
Assemble KM table and NCC data without duplicates when cohort data is
available
}
