% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{prep_km2}
\alias{prep_km2}
\title{Assemble KM table and NCC data without duplicates when cohort data is not
available}
\usage{
prep_km2(
  ncc,
  n_at_risk,
  id_name,
  set_id_name,
  t_name,
  t_match_name,
  y_name,
  match_var_names = NULL,
  n_per_case
)
}
\arguments{
\item{ncc}{(Matched) NCC data, if \code{cohort} is not available. A
\code{data.frame} or a matrix with column names. See Details below.}

\item{n_at_risk}{Number of subjects at risk at time of each cases in the NCC,
if \code{cohort} is not available. A \code{data.frame}, a matrix with
column names, or the path to the file that contains this information. See
Details below.}

\item{t_name}{Name of the variable in \code{cohort} for the time of event or
censoring. A \code{string}.}

\item{t_match_name}{Name of the time variable in \code{ncc} that is used to
match cases with controls. A \code{string}. See Details below.}

\item{y_name}{Name of the variable in \code{cohort} indicating the
event-censoring status of each subject, where 1 indicates events. A
\code{string}.}

\item{match_var_names}{Name(s) of the match variable(s) in \code{cohort} used
when drawing the NCC. A \code{string} vector. Default is \code{NULL}, i.e.,
the NCC was only time-matched.}

\item{n_per_case}{Number of controls matched to each case.}
}
\description{
Assemble KM table and NCC data without duplicates when cohort data is not
available
}
